split endpoints, simplify endpoints, catch no engine error elegantly
