+++ app.js
@@ -6,7 +6,8 @@ import v1Initialize from './routes/v1/initialize.js'
 import v1Engines from './routes/v1/engines.js'
 import v1EngineParameters from './routes/v1/engineParameters.js'
 import v1EngineGenerate from './routes/v1/engineGenerate.js'
-import v1Evals from './routes/v1/evals.js'
+import v1EvalsList from './routes/v1/evalsList.js'
+import v1EvalsTest from './routes/v1/evalsTest.js'
 import v1Leaderboard from './routes/v1/leaderboard.js'
 
 const app = express()
@@ -23,7 +24,8 @@ app.use("/api/v1/initialize", v1Initialize);
 app.use("/api/v1/engines", v1Engines);
 app.use("/api/v1/engines/", v1EngineParameters); //:engine/parameters
 app.use("/api/v1/engines/", v1EngineGenerate); //:engine/generate
-app.use("/api/v1/evals", v1Evals);
+app.use("/api/v1/evals", v1EvalsList);
+app.use("/api/v1/evals", v1EvalsTest);
 app.use("/api/v1/leaderboard", v1Leaderboard);
 
 app.listen(config.port, () => {
